moved movie dao to db folder
